pyproject.toml
[project]
name = "omphalOS"
version = "0.1.0"
description = "Contracts-first deterministic analytics workbench"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [{name = "Maintainers"}]
dependencies = [
  "pydantic>=2.0",
  "PyYAML>=6.0",
  "jsonschema>=4.0",
  "typer>=0.12",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.25",
  "duckdb>=0.10",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4",
  "mypy>=1.10",
  "types-PyYAML",
]

[project.scripts]
omphalos = "omphalos.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = ["src"]
